added some test, Order moved to the correct place
