+++ src/main/java/com/codecool/shop/controller/ProductControllerTest.java
@@ -1,0 +1,33 @@
+package com.codecool.shop.controller;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import javax.servlet.http.Cookie;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ProductControllerTest {
+
+    ProductController productController;
+
+    @BeforeEach
+    void setup() {
+        productController = new ProductController();
+    }
+
+    @Test
+    void doGet() {
+    }
+
+
+    @Test
+    void itemInCartAllNullTest() {
+        assertFalse(productController.itemInCart(null, null));
+    }
+
+    @Test
+    void itemInCartIdNullCookieEmpty() {
+        assertFalse(productController.itemInCart(null, new Cookie[0]));
+    }
+}
